+++ src/Swolfkrow/Start.cs
@@ -6,6 +6,30 @@
 public static partial class Workflow
 {
     /// <summary>
+    /// Starts an existing asynchronous workflow that yields a given sequence of events.
+    /// </summary>
+    /// <param name="events">An existing sequence of events.</param>
+    /// <typeparam name="TEvent">The (base) type of the events yielded by the resulting asynchronous workflow.</typeparam>
+    /// <returns>An asynchronous workflow that yields the given events in order.</returns>
+    /// <remarks>
+    /// This function serves as an entry point to the DSL.
+    /// </remarks>
+    public static IAsyncEnumerable<TEvent> Start<TEvent>(params TEvent[] events)
+        => Start((IEnumerable<TEvent>)events);
+
+    /// <summary>
+    /// Starts an existing asynchronous workflow that yields a given sequence of events.
+    /// </summary>
+    /// <param name="events">An existing sequence of events.</param>
+    /// <typeparam name="TEvent">The (base) type of the events yielded by the resulting asynchronous workflow.</typeparam>
+    /// <returns>An asynchronous workflow that yields the given events in order.</returns>
+    /// <remarks>
+    /// This function serves as an entry point to the DSL.
+    /// </remarks>
+    public static IAsyncEnumerable<TEvent> Start<TEvent>(IEnumerable<TEvent> events)
+        => events.ToAsyncEnumerable();
+
+    /// <summary>
     /// Starts an existing asynchronous workflow.
     /// </summary>
     /// <param name="workflow">An existing asynchronous workflow.</param>