+++ tests/Swolfkrow.Tests/Some.cs
@@ -10,6 +10,8 @@ internal static class Some
 {
     public record Event(string Description);
 
+    public record SpecificEvent(string Description) : Some.Event(Description);
+
     public static class Workflow
     {
         public static IAsyncEnumerable<Event> FromEvents(params Event[] events)